Relocate MapConverter
