+++ container-search/src/main/java/com/yahoo/search/dispatch/rpc/MapConverter.java
@@ -1,5 +1,5 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.searchlib.protobuf;
+package com.yahoo.search.dispatch.rpc;
 
 import ai.vespa.searchlib.searchprotocol.protobuf.Search.StringProperty;
 import ai.vespa.searchlib.searchprotocol.protobuf.Search.TensorProperty;
