--- searchlib/src/main/java/com/yahoo/searchlib/protobuf/package-info.java
@@ -1,5 +1,0 @@
-// Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-@ExportPackage
-package com.yahoo.searchlib.protobuf;
-
-import com.yahoo.osgi.annotation.ExportPackage;
